revert rename of types to _types
why: because types is imported/used by other services
