+++ wazo_auth_client/commands/tenants.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from wazo_lib_rest_client import RESTCommand
 
-from .._types import JSON
+from ..types import JSON
 
 
 class TenantsCommand(RESTCommand):
